Create plasma conftest file
